Re-read WiFi conf at first request
At the time of container start, it is assumed that the WiFi conf
available might not be the “correct” one: it can be changed
once the containers have started, when `ap` step is ran.
This thus re-reads the conf upon first /wifi/ request so it's using
the correct one.
Given requests only come via WiFi, it's not possible that this
would happen before the WiFi conf is applied.

Fixes #1